Move the participant list of the participant screen to sub-component
