+++ app/components/Participant/ParticipantMain.js
@@ -1,26 +1,14 @@
 import React from 'react';
-import { StyleSheet, ScrollView, TouchableOpacity } from 'react-native';
-import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
+import { StyleSheet, ScrollView } from 'react-native';
 
 import {LocalizationContext} from '../Translations';
 import ParticipantHeader from './ParticipantHeader';
-import ParticipantListItemInfo from '../Share/ParticipantListItemInfo';
+import ParticipantList from './ParticipantList';
 import EmptyListAction from '../Share/EmptyListAction';
 import AddNewParticipantMain from '../ParticipantModal/AddNewParticipantMain';
-import UncountableParticipantDetail from '../ParticipantModal/UncountableParticipantDetail';
 
-import { participantContentHeight, participantModalSnapPoints, uncountableParticipantDetailSnapPoints } from '../../constants/modal_constant';
+import { participantContentHeight, participantModalSnapPoints } from '../../constants/modal_constant';
 import { containerPaddingTop, containerPadding } from '../../utils/responsive_util';
-import Scorecard from '../../models/Scorecard';
-import Participant from '../../models/Participant';
-
-import Color from '../../themes/color';
-import listItemStyles from '../../themes/scorecardListItemStyle';
-import { getDeviceStyle } from '../../utils/responsive_util';
-import ParticipantListItemTabletStyles from '../../styles/tablet/ParticipantListItemComponentStyle';
-import ParticipantListItemMobileStyles from '../../styles/mobile/ParticipantListItemComponentStyle';
-
-const responsiveStyles = getDeviceStyle(ParticipantListItemTabletStyles, ParticipantListItemMobileStyles);
 
 class ParticipantMain extends React.Component {
   static contextType = LocalizationContext;
@@ -48,39 +36,14 @@ class ParticipantMain extends React.Component {
             addNewParticipant={() => this.showParticipantBottomSheet(null)} />
   }
 
-  showUncountedInfo() {
-    this.props.formModalRef.current?.setSnapPoints(uncountableParticipantDetailSnapPoints);
-    this.props.formModalRef.current?.setBodyContent(<UncountableParticipantDetail/>);
-    this.props.participantModalRef.current?.present();
-  }
-  
-  renderRightIcon(participant) {
-    if (!participant.countable)
-      return <TouchableOpacity onPress={() => this.showUncountedInfo()} style={responsiveStyles.uncountableButton}>
-                <Icon name='eye-off' color={Color.grayColor} size={24} />
-             </TouchableOpacity>
-  }
-
   renderParticipantList = () => {
-    const numberOfParticipant = Scorecard.find(this.props.scorecardUuid).number_of_participant;
-    this.totalParticipant = numberOfParticipant;
-    let doms = null;
-
-    if (Participant.getAll(this.props.scorecardUuid).length > 0) {
-      doms = this.props.participants.map((participant, index) =>
-        <ParticipantListItemInfo
-          key={index}
-          participant={participant}
-          onPress={() => this.showParticipantBottomSheet(participant)}
-          containerStyle={[responsiveStyles.itemContainer, listItemStyles.card]}
-          hasArrowIcon={true}
-          rightIcon={this.renderRightIcon(participant)}
-          arrowColor={Color.headerColor}
-        />
-      )
-    }
-
-    return doms;
+    return <ParticipantList
+              scorecardUuid={this.props.scorecardUuid}
+              participants={this.props.participants}
+              participantModalRef={this.props.participantModalRef}
+              formModalRef={this.props.formModalRef}
+              showParticipantBottomSheet={(participant) => this.showParticipantBottomSheet(participant)}
+           />
   }
 
   renderNoData() {
@@ -108,12 +71,7 @@ const styles = StyleSheet.create({
     paddingHorizontal: 14,
     flexGrow: 1,
     paddingTop: containerPaddingTop,
-  },
-  itemColumn: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
+  }
 });
 
 export default ParticipantMain;
